+++ src/datastore/async_methods/create.js
@@ -52,17 +52,10 @@ function create(resourceName, attrs, options) {
           _this.emit(definition, 'afterCreate', DSUtils.merge({}, attrs));
         }
         if (options.cacheResponse) {
-          var resource = _this.store[resourceName];
           var created = _this.inject(definition.name, attrs, options);
           var id = created[definition.idAttribute];
-          resource.completedQueries[id] = new Date().getTime();
-          resource.previousAttributes[id] = DSUtils.deepMixIn({}, created);
-          resource.saved[id] = DSUtils.updateTimestamp(resource.saved[id]);
-          resource.expiresHeap.push({
-            timestamp: resource.saved[id],
-            item: created
-          });
-          return _this.get(definition.name, id);
+          _this.store[resourceName].completedQueries[id] = new Date().getTime();
+          return created;
         } else {
           return _this.createInstance(resourceName, attrs, options);
         }