+++ src/datastore/sync_methods/inject.js
@@ -131,11 +131,6 @@ function _inject(definition, resource, attrs, options) {
           DSUtils.deepMixIn(resource.previousAttributes[id], attrs);
 
           resource.collection.push(item);
-          resource.expiresHeap.push({
-            item: item,
-            timestamp: DSUtils.updateTimestamp(resource.saved[id])
-          });
-
           resource.changeHistories[id] = [];
 
           if (DSUtils.w) {
@@ -166,6 +161,12 @@ function _inject(definition, resource, attrs, options) {
           }
         }
         resource.saved[id] = DSUtils.updateTimestamp(resource.saved[id]);
+        resource.expiresHeap.remove(item);
+        resource.expiresHeap.push({
+          item: item,
+          timestamp: resource.saved[id],
+          expires: definition.maxAge ? resource.saved[id] + definition.maxAge : Number.MAX_VALUE
+        });
         injected = item;
       } catch (err) {
         console.error(err.stack);